requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
Django==5.0.14
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
mypy-extensions==1.0.0
openai==1.76.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
